python-backend/requirements.txt
aistudio-sdk==0.3.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bce-python-sdk==0.9.46
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
cryptography==46.0.2
cssselect==1.3.0
cssutils==2.11.1
dotenv==0.9.9
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.19.1
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.27.0
google-auth==2.41.1
google-cloud-vision==3.11.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
imagesize==1.4.1
img2table==1.4.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.45.1
lxml==6.0.2
MarkupSafe==3.0.2
modelscope==1.30.0
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.0.1
networkx==3.5
ninja==1.13.0
numba==0.62.1
numpy==2.2.6
opencv-contrib-python==4.10.0.84
opencv-contrib-python-headless==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.2.0
paddlex==3.2.1
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
premailer==3.10.0
prettytable==3.16.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.10.23
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.9.30
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
xlsxwriter==3.2.9
